Add random delay in case of range collision
It's possible when adding 2 or more replica at the same time that one of the original instances returns a range, a new replica takes a part of that range and returns that during the query.
This would result in 2 ranges overlapping, and a `ValueError` would be thrown by `RangeSet#has` method.
To overcome this, the `ValueError` exception is catched and the query is retried after a random wait between 3-10 sec.
This should minimize the possibility of leaving an instance without id range.